pepper4/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.1.1
email-validator==2.0.0
python-dotenv==1.0.0 
Werkzeug==3.0.1
PyPDF2==3.0.1
python-magic==0.4.27
google-generativeai>=0.3.0
networkx>=3.1
matplotlib>=3.7.1
Pillow>=10.0.0
gunicorn
flask-migrate
flask_moment
pymysql
flask-cors
APScheduler
Flask-Mail
aiohttp
pytz
cloudinary
imapclient